docs: fix formula syntax
After fiddling for an hour, i gave up on ever getting anything close to
nicely readable. Either the canvas would eat fixes, or nothing got
fixed, or the root cause was never found (trailing double backslash?).

One could argue this is kind of unfair, but if the output contains Latex
resp. MathJAXX based formulas, they must be readable.

In the end i gave up and tried to understand the syntax and fix myself.